Update paths due to commands dir move
