+++ index.js
@@ -1,10 +1,14 @@
+const path = require('path');
 const fse = require('fs-extra');
 const lisky = require('vorpal')();
 const config = require('./config.json');
 
+const commandsDir = path.join(__dirname, 'src', 'commands');
+
 //import commands from ./commands/ folder
-fse.readdirSync('./commands').map(function (command) {
-	lisky.use(require('./commands/'+command));
+fse.readdirSync(commandsDir).forEach(function (command) {
+	const commandPath = path.join(commandsDir, command);
+	lisky.use(require(commandPath));
 });
 
 //Define vorpal
