+++ test/commands/get.js
@@ -2,7 +2,7 @@ const Vorpal = require('vorpal');
 const common = require('../common');
 const lisky = common.lisky;
 const sinon = common.sinon;
-const get = require('../../commands/get');
+const get = require('../../src/commands/get');
 const query = require('../../src/utils/query');
 const util = require('util');
