+++ test/utils/tablify.js
@@ -1,4 +1,4 @@
-const tablify = require('../../../src/utils/tablify');
+const tablify = require('../../src/utils/tablify');
 const expect = require('chai').expect;
 
 describe('#tablify', () => {
